requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
Pillow==10.2.0
pymongo==4.6.1
email-validator==2.1.0.post1